+++ README.md
@@ -62,3 +62,4 @@
           - 애노테이션은 리텐션@Retention 과 타겟@Target 을 무조건 있어야 한다
         - 합성 애노테이션
           - 반복적으로 애노테이션이 많이 붙는다면 합성 애노테이션을 붙혀서 코드를 좀 더 깔끔하게 할 수 있다.
+    2. 합성 애노테이션의 적용
