+++ src/main/java/study/spring/springboot/SpringbootApplication.java
@@ -1,26 +1,10 @@
 package study.spring.springboot;
 
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.web.embedded.tomcat.TomcatServletWebServerFactory;
-import org.springframework.boot.web.servlet.server.ServletWebServerFactory;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.DispatcherServlet;
+import study.spring.springboot.annotation.MySpringBootApplication;
 
-@Configuration
-@ComponentScan
+@MySpringBootApplication
 public class SpringbootApplication {
-
-    @Bean
-    public ServletWebServerFactory servletWebServerFactory() {
-        return new TomcatServletWebServerFactory();
-    }
-
-    @Bean
-    public DispatcherServlet dispatcherServlet() {
-        return new DispatcherServlet();
-    }
 
     public static void main(String[] args) {
         SpringApplication.run(SpringbootApplication.class, args);
